+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 
 * Revert adjusting bootstrap scripts for developing on Windows
 
+* Channel input - replaced ContentEditable with textarea
+
 [2.0.1-alpha.2]
 
 * UI layer for taken usernames for desktop and mobile