+++ CHANGELOG.md
@@ -12,6 +12,10 @@
 
 * Revert adjusting bootstrap scripts for developing on Windows
 
+* Channel input - replaced ContentEditable with textarea
+
+* Fix - up/down arrows now work properly inside channel input (textarea)
+
 [2.0.1-alpha.2]
 
 * UI layer for taken usernames for desktop and mobile